requirements.txt
aniso8601~=9.0.1
django==5.0.9
django-crispy-forms~=2.3
crispy-bootstrap3
django-formtools==2.5.1
django-model-utils==5.0.0
djangorestframework==3.15.2
django-debug-toolbar==4.3.0
django-dynamic-breadcrumbs==0.4.3
django-itemlist~=0.3.6
django-cas-ng~=5.0.1
django-shortuuidfield==0.1.3
django-proxy==1.1.0
django-taggit==0.22.2
django-schema-graph==3.1.0
pymemcache~=4.0.0
docutils==0.21.2
fuzzywuzzy~=0.18.0
ipaddress~=1.0.23
#ortools~=9.11.4210
lxml~=5.0.2
numpy~=2.1.3
scipy~=1.14.1
phonenumbers~=8.13.49
pycryptodome==3.21.0
python-dateutil~=2.9.0.post0
python-memcached==1.59
python-Levenshtein
psycopg2-binary
markdown
pytz~=2024.2
requests~=2.32.3
shortuuid
xhtml2pdf==0.2.16
pycountry==18.5.26
xlrd
unicodecsv
prettytable~=1.0.1